pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "agnes-ai-supply-chain-manager"
version = "0.1.0"
description = "RAG-powered multi-agent sourcing decision support system for CPG procurement"
requires-python = ">=3.11"
dependencies = [
    "google-cloud-bigquery>=3.11.0",
    "google-cloud-documentai>=2.20.0",
    "cognee>=0.1.0",
    "apify-client>=1.6.0",
    "playwright>=1.40.0",
    "elevenlabs>=1.0.0",
    "pulp>=2.7.0",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "hypothesis>=6.100.0",
    "jsonschema>=4.22.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
